+++ examples/webpack/main.js
@@ -1,0 +1,5 @@
+import star, { ReactComponent } from './star.url.svg'
+import Star from './star.simple.svg'
+
+console.log('url', typeof star, typeof ReactComponent)
+console.log('simple', typeof Star)