+++ src/cli/index.js
@@ -22,8 +22,9 @@ program
   .option('--no-svgo', 'disable SVGO')
   .option('--no-prettier', 'disable Prettier')
   .option('--template <file>', 'specify a custom template to use')
+  .option('--ext <ext>', 'specify a custom file extension (default: "js")')
   .option('--no-expand-props', 'disable props expanding')
-  .option('--ids', 'keep ids within the svg')
+  .option('--ref', 'add svgRef prop to svg')
   .option('--icon', 'use "1em" as width and height')
   .option('--no-dimensions', 'remove width and height')
   .option('--no-view-box', 'remove viewBox')
@@ -36,11 +37,15 @@ program
   .option(
     '-p, --precision <value>',
     'set the number of digits in the fractional part (svgo)',
+    parseInt,
   )
+  .option('--ids', 'keep ids within the svg (svgo)')
+  .option('--keep-useless-defs', 'keep elements of <defs> without id (svgo)')
   .option('--no-title', 'remove title tag (svgo)')
   .option(
-    '--tab-width',
+    '--tab-width <value>',
     'specify the number of spaces by indentation-level (prettier)',
+    parseInt,
   )
   .option('--use-tabs', 'indent lines with tabs instead of spaces (prettier)')
   .option('--no-semi', 'remove semi-colons (prettier)')