+++ src/cli/index.test.js
@@ -44,13 +44,6 @@ describe('cli', () => {
     expect(stdout).toMatchSnapshot()
   })
 
-  it('--no-dimensions', async () => {
-    const [stdout] = await exec(
-      'babel-node src/cli --no-dimensions __fixtures__/one.svg',
-    )
-    expect(stdout).toMatchSnapshot()
-  })
-
   it('--replace-attr-value', async () => {
     const [stdout] = await exec(
       'bin/svgr --replace-attr-value "#063855=currentColor" __fixtures__/one.svg',
@@ -59,7 +52,12 @@ describe('cli', () => {
   })
 
   it('--precision', async () => {
-    const [stdout] = await exec('bin/svgr --precision 1 __fixtures__/one.svg')
+    const [stdout] = await exec('bin/svgr --precision 10 __fixtures__/one.svg')
+    expect(stdout).toMatchSnapshot()
+  })
+
+  it('--tab-width', async () => {
+    const [stdout] = await exec('bin/svgr --tab-width 4 __fixtures__/one.svg')
     expect(stdout).toMatchSnapshot()
   })