+++ src/transforms/wrapIntoNativeComponent.js
@@ -16,6 +16,16 @@ export default (opts = {}) => (code, state) => {
     unsupportedComponents = new Set(),
   } = state
 
+  let props = ''
+
+  if (opts.expandProps && opts.ref) {
+    props = '{svgRef, ...props}'
+  } else if (opts.expandProps) {
+    props = 'props'
+  } else if (opts.ref) {
+    props = '{svgRef}'
+  }
+
   return `import React from 'react'
   import Svg, { ${componentsToList(
     reactNativeSvgReplacedComponents,
@@ -23,7 +33,7 @@ export default (opts = {}) => (code, state) => {
   ${logUnsupportedComponents(unsupportedComponents)}
 
 
-  const ${state.componentName} = (${opts.expandProps ? 'props' : ''}) => ${code}
+  const ${state.componentName} = (${props}) => ${code}
 
   export default ${state.componentName}`
 }